+++ src/components/Appointment.js
@@ -1,0 +1,8 @@
+import React from "react";
+import { Typography } from "@material-ui/core";
+
+function Appointment() {
+  return <Typography>Appointment details would go here</Typography>;
+}
+
+export default Appointment;
