+++ src/components/index.js
@@ -1,0 +1,6 @@
+import Appointment from "./Appointment";
+import CheckIn from "./CheckIn";
+import Welcome from "./Welcome";
+import Layout from "./Layout";
+
+export { Appointment, CheckIn, Welcome, Layout };
